tools/dx-agentic-dev-gen/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dx-agentic-dev-gen"
version = "0.1.0"
description = "DEEPX Agentic Development Platform Generator — generates platform-specific agent/skill/instruction files from canonical .deepx/ source"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
dx-agentic-gen = "dx_agentic_dev_gen.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
